python-agent-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
email-validator==2.1.0
python-dotenv==1.0.0
requests==2.31.0
azure-identity==1.15.0
azure-cognitiveservices-language-textanalytics==5.3.0
azure-ai-ml==1.12.0
azure-core==1.29.5
azure-ai-projects==1.0.0b4
openai==1.3.7
httpx==0.25.2
aiofiles==23.2.1
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
redis==5.0.1
celery==5.3.4
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0
rich==13.7.0
structlog==23.2.0